Reorg: Rename `mod channel` to `mod mnt`
All the code for performing mounts/unmounts is in `mod mnt`, but not much
else.  The external API is limited to `Mount` and `MountOptions`.

`Channel` exists seperately and doesn't depend on anything from `mnt`.  It
contains no conditional compilation.
